NAVS refactored.
