+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/composite/section/navs/BaseTests.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Test;
 import static io.github.com.StaticSite.bsPage;
 import static io.github.com.pages.BootstrapPage.navsBaseA;
 import static io.github.com.pages.BootstrapPage.navsBaseLi;
-import static io.github.epam.bootstrap.tests.composite.section.navs.ClickVarification.testElementClick;
+import static io.github.epam.bootstrap.tests.composite.section.navs.ClickVerification.testElementClick;
 import static io.github.epam.states.States.shouldBeLoggedIn;
 import static org.hamcrest.CoreMatchers.is;
