+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/composite/section/navs/PillsTests.java
@@ -8,7 +8,7 @@ import org.testng.annotations.Test;
 
 import static io.github.com.StaticSite.bsPage;
 import static io.github.com.pages.BootstrapPage.navsPills;
-import static io.github.epam.bootstrap.tests.composite.section.navs.ClickVarification.testElementClick;
+import static io.github.epam.bootstrap.tests.composite.section.navs.ClickVerification.testElementClick;
 import static io.github.epam.states.States.shouldBeLoggedIn;
 import static org.hamcrest.CoreMatchers.is;
